requirements.txt
qdrant-client
openai
langchain
langchain-openai
tqdm
python-dotenv
pandas